+++ examples/schema.rs
@@ -1,5 +1,5 @@
 use rmpv::Value;
-use tarantool_rs::{Connection, ExecutorExt, IteratorType};
+use tarantool_rs::{Connection, Executor, ExecutorExt, IteratorType};
 use tracing::info;
 
 #[tokio::main]
@@ -7,8 +7,9 @@ async fn main() -> Result<(), anyhow::Error> {
     pretty_env_logger::init();
 
     let conn = Connection::builder().build("127.0.0.1:3301").await?;
+    let tx = conn.transaction().await?;
 
-    let data = conn.space("clients").await?;
+    let data = tx.space("clients").await?;
     info!("{:?}", data);
     let space = data.unwrap();
     info!(
@@ -17,13 +18,21 @@ async fn main() -> Result<(), anyhow::Error> {
             .select::<(i64, String), _>(None, None, Some(IteratorType::All), ())
             .await?
     );
-    space.upsert((0, "Name"), ("=",)).await?;
-    space
-        .update(
-            (0,),
-            (Value::Array(vec!["=".into(), 1.into(), "Second".into()]),),
-        )
-        .await?;
+    info!("UPSERT: {:?}", space.upsert((0, "Name"), ("=",)).await?);
+    info!(
+        "UPDATE: {:?}",
+        space
+            .update(
+                (0,),
+                (rmpv::Value::Array(vec![
+                    "=".into(),
+                    1.into(),
+                    "Second".into()
+                ]),),
+            )
+            .await?
+    );
+    info!("DELETE: {:?}", space.delete((2,)).await?);
     info!(
         "Post: {:?}",
         space