+++ src/client/executor_ext.rs
@@ -12,7 +12,7 @@ use crate::{
     schema::{SchemaEntityKey, Space},
     tuple::Tuple,
     utils::extract_and_deserialize_iproto_data,
-    IteratorType, Result, SqlResponse, TupleResponse,
+    CallResponse, DmoResponse, IteratorType, Result, SqlResponse,
 };
 
 /// Helper trait around [`Executor`] trait, which allows to send specific requests
@@ -37,12 +37,12 @@ pub trait ExecutorExt: Executor {
     /// Evaluate Lua expression.
     ///
     /// Check [docs][crate#deserializing-lua-responses-in-call-and-eval] on how to deserialize response.
-    async fn eval<A, I>(&self, expr: I, args: A) -> Result<TupleResponse>
+    async fn eval<A, I>(&self, expr: I, args: A) -> Result<CallResponse>
     where
         A: Tuple + Send,
         I: AsRef<str> + Send + Sync,
     {
-        Ok(TupleResponse(
+        Ok(CallResponse(
             self.send_request(Eval::new(expr.as_ref(), args)).await?,
         ))
     }
@@ -50,12 +50,12 @@ pub trait ExecutorExt: Executor {
     /// Remotely call function in Tarantool.
     ///
     /// Check [docs][crate#deserializing-lua-responses-in-call-and-eval] on how to deserialize response.
-    async fn call<A, I>(&self, function_name: I, args: A) -> Result<TupleResponse>
+    async fn call<A, I>(&self, function_name: I, args: A) -> Result<CallResponse>
     where
         A: Tuple + Send,
         I: AsRef<str> + Send + Sync,
     {
-        Ok(TupleResponse(
+        Ok(CallResponse(
             self.send_request(Call::new(function_name.as_ref(), args))
                 .await?,
         ))
@@ -83,63 +83,67 @@ pub trait ExecutorExt: Executor {
         extract_and_deserialize_iproto_data(body).map_err(Into::into)
     }
 
-    // TODO: decode response
     /// Insert tuple.
-    async fn insert<T>(&self, space_id: u32, tuple: T) -> Result<()>
+    async fn insert<T>(&self, space_id: u32, tuple: T) -> Result<DmoResponse>
     where
         T: Tuple + Send,
     {
-        let _ = self.send_request(Insert::new(space_id, tuple)).await?;
-        Ok(())
+        Ok(DmoResponse(
+            self.send_request(Insert::new(space_id, tuple)).await?,
+        ))
     }
 
-    // TODO: decode response
     /// Update tuple.
-    async fn update<K, O>(&self, space_id: u32, index_id: u32, keys: K, ops: O) -> Result<()>
+    async fn update<K, O>(
+        &self,
+        space_id: u32,
+        index_id: u32,
+        keys: K,
+        ops: O,
+    ) -> Result<DmoResponse>
     where
         K: Tuple + Send,
         O: Tuple + Send,
     {
-        let _ = self
-            .send_request(Update::new(space_id, index_id, keys, ops))
-            .await?;
-        Ok(())
+        Ok(DmoResponse(
+            self.send_request(Update::new(space_id, index_id, keys, ops))
+                .await?,
+        ))
     }
 
-    // TODO: decode response
     /// Update or insert tuple.
-    async fn upsert<T, O>(&self, space_id: u32, tuple: T, ops: O) -> Result<()>
+    async fn upsert<T, O>(&self, space_id: u32, tuple: T, ops: O) -> Result<DmoResponse>
     where
         T: Tuple + Send,
         O: Tuple + Send,
     {
-        let _ = self.send_request(Upsert::new(space_id, ops, tuple)).await?;
-        Ok(())
+        Ok(DmoResponse(
+            self.send_request(Upsert::new(space_id, ops, tuple)).await?,
+        ))
     }
 
     // TODO: structured tuple
-    // TODO: decode response
     /// Insert a tuple into a space. If a tuple with the same primary key already exists,
     /// replaces the existing tuple with a new one.
-    async fn replace<T>(&self, space_id: u32, tuple: T) -> Result<()>
+    async fn replace<T>(&self, space_id: u32, tuple: T) -> Result<DmoResponse>
     where
         T: Tuple + Send,
     {
-        let _ = self.send_request(Replace::new(space_id, tuple)).await?;
-        Ok(())
+        Ok(DmoResponse(
+            self.send_request(Replace::new(space_id, tuple)).await?,
+        ))
     }
 
     // TODO: structured tuple
-    // TODO: decode response
     /// Delete a tuple identified by the primary key.
-    async fn delete<T>(&self, space_id: u32, index_id: u32, keys: T) -> Result<()>
+    async fn delete<T>(&self, space_id: u32, index_id: u32, keys: T) -> Result<DmoResponse>
     where
         T: Tuple + Send,
     {
-        let _ = self
-            .send_request(Delete::new(space_id, index_id, keys))
-            .await?;
-        Ok(())
+        Ok(DmoResponse(
+            self.send_request(Delete::new(space_id, index_id, keys))
+                .await?,
+        ))
     }
 
     // TODO: options