+++ src/client/mod.rs
@@ -1,24 +1,26 @@
 pub use self::{
+    call_response::CallResponse,
     connection::Connection,
+    dmo_response::DmoResponse,
     executor::Executor,
     executor_ext::ExecutorExt,
     prepared_sql_statement::PreparedSqlStatement,
     sql_response::SqlResponse,
     stream::Stream,
     transaction::{Transaction, TransactionBuilder},
-    tuple_response::TupleResponse,
 };
 
 pub mod schema;
 
+mod call_response;
 mod connection;
+mod dmo_response;
 mod executor;
 mod executor_ext;
 mod prepared_sql_statement;
 mod sql_response;
 mod stream;
 mod transaction;
-mod tuple_response;
 
 mod private {
     use crate::client::{Connection, Stream, Transaction};