+++ tests/test_data.lua
@@ -4,6 +4,8 @@ box.cfg {}
 box.schema.user.create('Sisko', {password = 'A-4-7-1'})
 
 -- Table with test data
+box.schema.sequence.create('seq_id_ds9_crew',{min=1, start=1})
+
 ds9_crew = box.schema.space.create('ds9_crew')
 ds9_crew:format({
       {'id', type = 'unsigned'},
@@ -11,11 +13,12 @@ ds9_crew:format({
       {'rank', type = 'string', is_nullable = true},
       {'occupation', type = 'string', is_nullable = true},
 })
-ds9_crew:create_index('idx_id', {parts = {1, 'unsigned'}})
+ds9_crew:create_index('idx_id', {parts = {1, 'unsigned'}, sequence = 'seq_id_ds9_crew'})
 ds9_crew:create_index('idx_name', {parts = {2, 'string'}})
 ds9_crew:create_index('idx_rank', {unique = false, parts = {3, 'string', is_nullable = true}})
 
-ds9_crew:auto_increment{'Benjamin Sisko', 'Captain', 'Commanding officer'}
+
+ds9_crew:auto_increment{'Benjamin Sisko', 'Commander', 'Commanding officer'}
 ds9_crew:auto_increment{'Kira Nerys', 'Major', 'First officer'}
 ds9_crew:auto_increment{'Jadzia Dax', 'Lieutenant Commander', 'Science officer'}
 ds9_crew:auto_increment{'Julian Bashir', 'Lieutenant', 'Chief medical officer'}