+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.8] - 2023-09-04
+### Added
+ - Data-manipulation operations (insert, update, upsert, replace, delete) now return `DmoResponse` with row, returned by operation (#7).
+
+### Changed
+ - `TupleResponse` renamed to `CallResponse`.
+
+
 ## [0.0.7] - 2023-08-24
 ### Added
  - Support for preparing and executing SQL queries.